Changes to ErlangStatus, removing twitter/Weekly
* Rebranding to ErlangStatus

Renamed ElixirStatus to ErlangStatus in main html pages,
also rephrased a few sentences based on the Rust version of this page:
https://herald.community.rs

* Removed Twitter and ElixirWeekly mentions

As we do not have these two resources for now, commented out so it can
easily be put back in once we do.

* Placeholder logo

Created a placeholder logo based on the Erlang logo, only changed colors
and removed writing because i'm not sure we can use the actual
Erlang logo.